last commit before merge attempt

+++ server.R
@@ -4,9 +4,14 @@ library("readr")
 library("dplyr")
 
 write_date <- function(){
-        date() %>%
-                as.data.frame() %>%
+        if (exists("TEST")) {
+                read_csv("TEST") %>%
+                rbind(date()) %>%
                 write_csv("TEST")
+        }
+        else {
+                write_csv(as.data.frame(date()), "TEST")
+        }       
 }
 
 shinyServer(function(input, output, session) {
